C++ syntax
I have now understood the following terms and their meaning:

Class: A template/blueprint that describes the states and, or behaviours that objects of its type support.  For example, "Canine" is a class of animal for which dogs (object) belong

Object: Objects have states and behaviours are an instance of a class. For example a dog can be an object of the canine class.  The dog has states - i.e. physical attributes, and behaviours - i.e. barking, eating etc.

Methods: Methods are basically class or object behaviours.  Action taken is executed by writing and manipulating the data in the methods

Instance variable: An object’s states are defined by the values of an instance variable.  Objects have many instance variables.